% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Binding Factors together}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factor
}
\description{
Create a new factor, made of the 2 previously seperate factors
by a union of their contents and structure.
}
\examples{
fbind(c("one", "two", "3"), 3:5)
}
