% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praise.R
\name{praise}
\alias{praise}
\title{Almost a compliment}
\usage{
praise(name)
}
\arguments{
\item{name}{text string: Name of person to compliment.}
}
\value{
text string
}
\description{
Help to make someone feel better, but not too good.
}
\examples{
function(name = "Jojo")
}
